Move getResourcesDataByGroups method to utils
